Add support for non-struct pointers (#5)
This PR adds support for pointers to optional fields in forms.

This example fails in the old version:
```go
type TestForm struct {
	Test string `binding:"Url"`
	TestPtr *string `binding:"Url"`
}

empty := ""
form := &TestForm{
	Test: "",
	TestPtr: &empty,
}
_ := RawValidate(form)
```
The empty `Test` field is no problem because an empty field is valid url. The empty pointer field throws an error because the regex checks against the wrong field value of `"0xc000011dc0"`. So the memory address is formatted as string and tested. After the change both fields pass because both are treated equal.

Fixes the workaround in https://github.com/go-gitea/gitea/pull/15690

Co-authored-by: Lunny Xiao <xiaolunwen@gmail.com>
Reviewed-on: https://gitea.com/go-chi/binding/pulls/5
Reviewed-by: Lunny Xiao <xiaolunwen@gmail.com>
Co-authored-by: KN4CK3R <admin@oldschoolhack.me>
Co-committed-by: KN4CK3R <admin@oldschoolhack.me>